+++ packages/terra-embedded-content-consumer/src/EmbeddedContentConsumer.jsx
@@ -8,6 +8,29 @@ const propTypes = {
    */
   src: PropTypes.string.isRequired,
   /**
+   * The inline HTML content to render within the iframe.
+   *
+   * When `srcdoc` prop is set, this will override the `src` URL.
+   * It's important to know that when this option is used,
+   * the [xfc](https://github.com/cerner/xfc) library does not manage
+   * [auto resizing](https://github.com/cerner/xfc#iframe-resizing-config)
+   * of the content of the iframe even if `resizeConfig` option is set. It's
+   * important to specify the width and height of the content.
+   *
+   * To set the width and height of the iframe, use the `iframeAttrs` config.
+   *
+   * ```
+   *  options={{
+   *     iframeAttrs:
+   *     {
+   *       width: '100%', // width of the iframe
+   *       height: '400px', // height of the iframe
+   *     },
+   *   }}
+   * ```
+   */
+  srcdoc: PropTypes.string,
+  /**
    * Notifies the component that the container has been mounted. Provides a reference
    * to this component to allow triggering messages on the embedded application.
    */
@@ -22,7 +45,22 @@ const propTypes = {
   onAuthorize: PropTypes.func,
   /**
    * The component can be configured with consumer frame options.
+   *
+   * Example `options` object:
+   *
+   * ```
+   * options = {{
+   *   iframeAttrs: {
+   *     id: 'iframe-id', // Sets iframe id
+   *     width: '100%', // Optional - Sets the width to 100%
+   *     height: '400px', // Optional - Sets the height to 400px; must not be 100%
+   *     frameborder: '0', // Optional - '0' Disable frame border, '1' Sets frame border
+   *   },
+   * }}
+   * ```
+   *
    * See xfc consumer configuration for details: https://github.com/cerner/xfc
+   *
    */
   options: PropTypes.object,
   /**
@@ -37,8 +75,23 @@ const propTypes = {
 
 class EmbeddedContentConsumer extends React.Component {
   componentDidMount() {
+    // Merging the iframe options props
+    const frameOptions = { ...this.props.options };
+    if (frameOptions.iframeAttrs == null) {
+      frameOptions.iframeAttrs = {};
+    }
+
+    if (this.props.srcdoc) {
+      Object.assign(frameOptions.iframeAttrs, { srcdoc: this.props.srcdoc });
+      Object.keys(frameOptions.iframeAttrs).forEach(key => {
+        if (frameOptions.iframeAttrs[key] == null) {
+          delete frameOptions.iframeAttrs[key];
+        }
+      });
+    }
+
     // Mount the provided source as the application into the content wrapper.
-    this.xfcFrame = Consumer.mount(this.embeddedContentWrapper, this.props.src, this.props.options);
+    this.xfcFrame = Consumer.mount(this.embeddedContentWrapper, this.props.src, frameOptions);
 
     // Notify that the consumer frame has mounted.
     if (this.props.onMount) {
@@ -66,6 +119,7 @@ class EmbeddedContentConsumer extends React.Component {
   render() {
     const {
       src,
+      srcdoc,
       onMount,
       onLaunch,
       onAuthorize,